+++ src/main/java/com/almostreliable/lootjs/core/entry/EmptyLootEntry.java
@@ -1,6 +1,8 @@
 package com.almostreliable.lootjs.core.entry;
 
 import com.almostreliable.lootjs.util.DebugInfo;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraft.world.level.storage.loot.entries.EmptyLootItem;
 import net.minecraft.world.level.storage.loot.entries.LootPoolSingletonContainer;
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
@@ -28,5 +30,10 @@ public class EmptyLootEntry extends AbstractSimpleLootEntry<EmptyLootItem> imple
     public void collectDebugInfo(DebugInfo info) {
         info.add("% Empty");
         super.collectDebugInfo(info);
+    }
+
+    @Override
+    public ItemStack createRandomItem(LootContext context) {
+        return ItemStack.EMPTY;
     }
 }