+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning].
 ## Unreleased
 
 - Fix `addCondition` not being visible for loot condition containers
+- Allow composite loot entries like alternatives, sequences and groups in `replaceLoot` action
 
 ## [3.6.0] - 2025-10-14